+++ src/core/datastructures/data.cpp
@@ -23,7 +23,7 @@ void Data::copyRepresentations( Data* targetData ) const{
 void Data::addRepresentation(DataRepresentation* representation) {
     representations_.push_back(representation);
     lastValidRepresentation_ = representation;
-    setRepresentationAsValid(representations_.size()-1);
+    setRepresentationAsValid(static_cast<int>(representations_.size())-1);
 }
 
 bool Data::hasRepresentations() const {
